12/requirements.txt
celery==5.3.6
Django==4.2.2
djangorestframework==3.14.0
requests==2.31.0
psycopg2==2.9.6
django-celery-beat==2.5.0
django-celery-results==2.5.1
redis==5.0.0
